Price now drops as items are removed. However the price can blow up after adding to many items due to being a double. So I will be working on making the total price stable.

+++ MenuApp/Models/Entity Framework/Order.cs
@@ -19,11 +19,12 @@ namespace MenuApp.Models
         public string customerName { get; set; }
 
         [Required]
-        //   public List<Item> orderItems { get; set; }
+       // public List<itemFromJson> orderItems { get; set; }
 
         public string orderItems { get; set; }
-        public int orderPrice { get; set; }
+        public double orderPrice { get; set; }
         public DateTime timeOfOrder { get; set; }
 
     }
+
 }
